Moved preference keyboard on start in UX forwarder
